+++ src/components/Movies/Movies.css
@@ -7,6 +7,7 @@
 
 .movie__button-wrapper {
   display: flex;
+  padding-bottom: 80px;
 }
 
 .movie__open-more {
