+++ src/components/SearchForm/SearchForm.js
@@ -1,35 +1,11 @@
-import { useState } from 'react';
 import FilterCheckbox from '../FilterCheckbox/FilterCheckbox';
-import * as movieApi from '../../utils/MovieApi';
 
-function SearchForm() {
-  const [movieSearch, setMovieSearch] = useState('');
-  const [movieSearchError, setMovieSearchError] = useState('');
-  const [movieList, setMovieList] = useState('');
-
-  const handleSubmit = (e) => {
-    const arr = [];
-    e.preventDefault();
-    if (movieSearch.length === 0) {
-      setMovieSearchError('Нужно ввести ключевое слово');
-    } else {
-      movieApi
-        .getMovies()
-        .then((res) =>
-          res.forEach((item) => {
-            arr.push(item);
-          })
-        )
-        .then((res) => localStorage.setItem('movieList', JSON.stringify(arr)))
-        .catch((err) => console.log(`${err.status}: ${err.message}`));
-      setMovieSearchError('');
-    }
-  };
-
-  const handleMovieInput = (e) => {
-    setMovieSearch(e.target.value);
-  };
-
+function SearchForm({
+  handleSubmit,
+  handleMovieInput,
+  movieSearch,
+  movieSearchError,
+}) {
   return (
     <>
       <section className='search'>
